requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.3
idna==2.8
paho-mqtt==1.4.0
PyJWT==1.7.1
PySocks==1.6.8
python-decouple==3.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
twilio==6.26.3
urllib3==1.24.2
